+++ built_value_generator/test/built_value_generator_test.dart
@@ -105,7 +105,7 @@ class Value implements Built<Value, ValueBuilder> {
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
-}'''), contains("1. Make class abstract."));
+}'''), contains('1. Make class abstract.'));
     });
 
     test('suggests correct Built type parameters for implements', () async {
@@ -265,6 +265,114 @@ abstract class Value implements Built<Value, ValueBuilder> {
 }'''), isNot(contains('1.')));
     });
 
+    test('suggests making builder initializer static', () async {
+      expect(
+          await generate('''library value;
+import 'package:built_value/built_value.dart';
+part 'value.g.dart';
+abstract class Value implements Built<Value, ValueBuilder> {
+  void _initializeBuilder(ValueBuilder b) {}
+  Value._();
+  factory Value([void Function(ValueBuilder) updates]) = _\$Value;
+}
+abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
+  ValueBuilder._();
+  factory ValueBuilder() = _\$ValueBuilder;
+}'''),
+          contains('1. Fix _initializeBuilder signature: '
+              'static void _initializeBuilder(ValueBuilder b)'));
+    });
+
+    test('suggests making builder initializer return void', () async {
+      expect(
+          await generate('''library value;
+import 'package:built_value/built_value.dart';
+part 'value.g.dart';
+abstract class Value implements Built<Value, ValueBuilder> {
+  static String _initializeBuilder(ValueBuilder b) {}
+  Value._();
+  factory Value([void Function(ValueBuilder) updates]) = _\$Value;
+}
+abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
+  ValueBuilder._();
+  factory ValueBuilder() = _\$ValueBuilder;
+}'''),
+          contains('1. Fix _initializeBuilder signature: '
+              'static void _initializeBuilder(ValueBuilder b)'));
+    });
+
+    test('suggests making builder initializer accept a builder', () async {
+      expect(
+          await generate('''library value;
+import 'package:built_value/built_value.dart';
+part 'value.g.dart';
+abstract class Value implements Built<Value, ValueBuilder> {
+  static void _initializeBuilder(String b) {}
+  Value._();
+  factory Value([void Function(ValueBuilder) updates]) = _\$Value;
+}
+abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
+  ValueBuilder._();
+  factory ValueBuilder() = _\$ValueBuilder;
+}'''),
+          contains('1. Fix _initializeBuilder signature: '
+              'static void _initializeBuilder(ValueBuilder b)'));
+    });
+
+    test('suggests making builder finalizer static', () async {
+      expect(
+          await generate('''library value;
+import 'package:built_value/built_value.dart';
+part 'value.g.dart';
+abstract class Value implements Built<Value, ValueBuilder> {
+  void _finalizeBuilder(ValueBuilder b) {}
+  Value._();
+  factory Value([void Function(ValueBuilder) updates]) = _\$Value;
+}
+abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
+  ValueBuilder._();
+  factory ValueBuilder() = _\$ValueBuilder;
+}'''),
+          contains('1. Fix _finalizeBuilder signature: '
+              'static void _finalizeBuilder(ValueBuilder b)'));
+    });
+
+    test('suggests making builder finalizer return void', () async {
+      expect(
+          await generate('''library value;
+import 'package:built_value/built_value.dart';
+part 'value.g.dart';
+abstract class Value implements Built<Value, ValueBuilder> {
+  static String _finalizeBuilder(ValueBuilder b) {}
+  Value._();
+  factory Value([void Function(ValueBuilder) updates]) = _\$Value;
+}
+abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
+  ValueBuilder._();
+  factory ValueBuilder() = _\$ValueBuilder;
+}'''),
+          contains('1. Fix _finalizeBuilder signature: '
+              'static void _finalizeBuilder(ValueBuilder b)'));
+    });
+
+    test('suggests making builder finalizer accept a builder', () async {
+      expect(
+          await generate('''library value;
+import 'package:built_value/built_value.dart';
+part 'value.g.dart';
+abstract class Value implements Built<Value, ValueBuilder> {
+  static void _finalizeBuilder(String b) {}
+  Value._();
+  factory Value([void Function(ValueBuilder) updates]) = _\$Value;
+}
+abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
+  ValueBuilder._();
+  factory ValueBuilder() = _\$ValueBuilder;
+}'''),
+          contains('1. Fix _finalizeBuilder signature: '
+              'static void _finalizeBuilder(ValueBuilder b)'));
+    });
+
     test('suggests to add constructor to value class', () async {
       expect(await generate('''library value;
 import 'package:built_value/built_value.dart';
@@ -275,7 +383,7 @@ abstract class Value implements Built<Value, ValueBuilder> {
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
-}'''), contains("1. Make class have exactly one constructor: Value._();"));
+}'''), contains('1. Make class have exactly one constructor: Value._();'));
     });
 
     test('suggests to add constructor when there is synthetic constructor',
@@ -288,7 +396,7 @@ abstract class Value implements Built<Value, ValueBuilder> {
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
-}'''), contains("1. Make class have exactly one constructor: Value._();"));
+}'''), contains('1. Make class have exactly one constructor: Value._();'));
     });
 
     test('allows code in constructor of value class', () async {
@@ -307,7 +415,7 @@ abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   factory ValueBuilder() = _\$ValueBuilder;
 }'''),
           isNot(contains(
-              "1. Make class have exactly one constructor: Value._();")));
+              '1. Make class have exactly one constructor: Value._();')));
     });
 
     test('suggests to remove constructor from non-instantiable value class',
@@ -370,7 +478,7 @@ abstract class Value implements Built<Value, ValueBuilder> {
 class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
-}'''), contains("1. Make builder class abstract"));
+}'''), contains('1. Make builder class abstract'));
     });
 
     test('suggests correct Builder type parameters', () async {
@@ -403,8 +511,8 @@ abstract class Value implements Built<Value, ValueBuilder> {
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   factory ValueBuilder() = _\$ValueBuilder;
 }'''),
-          contains("1. Make builder class "
-              "have exactly one constructor: ValueBuilder._();"));
+          contains('1. Make builder class '
+              'have exactly one constructor: ValueBuilder._();'));
     });
 
     test('suggests constructor for builder class with synthetic constructor',
@@ -419,8 +527,8 @@ abstract class Value implements Built<Value, ValueBuilder> {
 }
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
 }'''),
-          contains("1. Make builder class "
-              "have exactly one constructor: ValueBuilder._();"));
+          contains('1. Make builder class '
+              'have exactly one constructor: ValueBuilder._();'));
     });
 
     test('suggests to add factory to builder class', () async {
@@ -435,8 +543,8 @@ abstract class Value implements Built<Value, ValueBuilder> {
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
 }'''),
-          contains("1. Make builder class have exactly one factory: "
-              "factory ValueBuilder() = _\$ValueBuilder;"));
+          contains('1. Make builder class have exactly one factory: '
+              'factory ValueBuilder() = _\$ValueBuilder;'));
     });
 
     test('suggests value fields must be getters', () async {
@@ -452,7 +560,7 @@ abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
   int foo;
-}'''), contains("1. Make field foo a getter."));
+}'''), contains('1. Make field foo a getter.'));
     });
 
     test('suggests value fields must be public', () async {
@@ -511,7 +619,7 @@ abstract class Value implements Built<Value, ValueBuilder> {
 abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
-}'''), contains("1. Make builder have exactly these fields: foo"));
+}'''), contains('1. Make builder have exactly these fields: foo'));
     });
 
     test('suggests builder fields must be same type', () async {
@@ -527,7 +635,7 @@ abstract class ValueBuilder implements Builder<Value, ValueBuilder> {
   ValueBuilder._();
   factory ValueBuilder() = _\$ValueBuilder;
   String foo;
-}'''), contains("1. Make builder field foo have type: int"));
+}'''), contains('1. Make builder field foo have type: int'));
     });
 
     test('ignores setters', () async {