+++ end_to_end_test/lib/polymorphism.dart
@@ -19,9 +19,18 @@ abstract class Animal extends Object with Walker implements OtherInterface {
 // still being allowed to implement any other interface.
 abstract class OtherInterface {}
 
+// Check a second layer of `instantiable: false` inheritance.
+@BuiltValue(instantiable: false)
+abstract class Mammal implements Animal {
+  @override
+  Mammal rebuild(void Function(MammalBuilder) updates);
+  @override
+  MammalBuilder toBuilder();
+}
+
 abstract class Cat extends Object
     with Walker
-    implements Animal, Built<Cat, CatBuilder> {
+    implements Mammal, Built<Cat, CatBuilder> {
   static Serializer<Cat> get serializer => _$catSerializer;
 
   bool get tail;
@@ -102,7 +111,7 @@ abstract class StandardCat extends Object
 abstract class HasField<T> {
   T get field;
 
-  HasField<T> rebuild(void updates(HasFieldBuilder<T> b));
+  HasField<T> rebuild(void Function(HasFieldBuilder<T>) updates);
   HasFieldBuilder<T> toBuilder();
 }